requirements.txt
google-genai
yt-dlp
static-ffmpeg
faster-whisper
